refactor(modal-dialog): Extract tabs into separate components
Creates the following components of each tab:
- user-modal-sidebar
- user-modal-tab-absences
- user-modal-tab-profiles
- user-modal-tab-timesheet
