+++ src/scenes/user-modal/user-modal-tab-absences.jsx
@@ -1,0 +1,37 @@
+// React API
+import * as React from 'react';
+
+// External components
+import Box from '@mui/material/Box';
+
+// In-house components
+
+// API
+
+// Icons
+
+// Images imports
+
+// Styles imports
+
+// Style Objects
+
+// Create a FormikContext
+
+// Transitions
+
+// ------------------------
+// Absences Tab component
+// ------------------------
+const UserModalTabAbsences = (props) =>  {
+
+    if (props.tabExpanded === 'Absences') {
+      return(
+        <Box>
+            This is "Absences" tab
+        </Box>
+      )
+    }
+}
+
+export default UserModalTabAbsences;
