+++ src/scenes/user-modal/user-modal-tab-timesheet.jsx
@@ -1,0 +1,37 @@
+// React API
+import * as React from 'react';
+
+// External components
+import Box from '@mui/material/Box';
+
+// In-house components
+
+// API
+
+// Icons
+
+// Images imports
+
+// Styles imports
+
+// Style Objects
+
+// Create a FormikContext
+
+// Transitions
+
+// ------------------------
+// Timesheet Tab component
+// ------------------------
+const UserModalTabTimesheet = (props) =>  {
+
+    if (props.tabExpanded === 'Timesheet') {
+      return(
+        <Box>
+            This is "Timesheet" tab
+        </Box>
+      )
+    }
+}
+
+export default UserModalTabTimesheet;
